task 4.1 (SQL-requests) - new AvatarController

+++ src/main/java/ru/hogwarts/school/repositories/StudentRepository.java
@@ -18,7 +18,7 @@ public interface StudentRepository extends JpaRepository<Student, Long> {
     int getNumberOfAllStudents();
 
     @Query(value = "SELECT avg(age) AS AverageAge FROM student", nativeQuery = true)
-    int getAverageAgeOfStudents();
+    double getAverageAgeOfStudents();
 
     @Query(value = "SELECT * FROM student ORDER BY id DESC LIMIT 5", nativeQuery = true)
     List<Student> getLast5Students();
